[160_common_interface], issue #160, defined common OxygenSolubilityInterface
