+++ lbh15/properties/lbe_thermochemical_properties/solubility_in_lbe.py
@@ -4,10 +4,10 @@ from typing import List
 from typing import Union
 import numpy as np
 from scipy.constants import atm
-from lbh15.properties.interface import PropertyInterface
 from lbh15.properties.common_interface import IronSolubilityInterface
 from lbh15.properties.common_interface import NickelSolubilityInterface
 from lbh15.properties.common_interface import ChromiumSolubilityInterface
+from lbh15.properties.common_interface import OxygenSolubilityInterface
 from ..._decorators import range_warning
 
 
@@ -413,7 +413,7 @@ class ChromiumSolubilityMartynov1998(LBEChromiumSolubilityInterface):
         return [673.0, 773.0]
 
 
-class OxygenSolubility(PropertyInterface):
+class OxygenSolubility(OxygenSolubilityInterface):
     """
     Liquid lbe *Oxygen solubility* property class.
     """
@@ -443,33 +443,12 @@ class OxygenSolubility(PropertyInterface):
         return np.power(10, 2.25-4125/T)
 
     @property
-    def name(self) -> str:
-        """
-        str : Name of the property
-        """
-        return "o_sol"
-
-    @property
-    def units(self) -> str:
-        """
-        str : Oxygen solubility unit
-        """
-        return "[wt.%]"
-
-    @property
     def range(self) -> List[float]:
         """
         List[float] : Temperature validity range of the Oxygen solubility
         correlation function
         """
         return [673.0, 1013.0]
-
-    @property
-    def long_name(self) -> str:
-        """
-        str : Oxygen solubility long name
-        """
-        return "oxygen solubility"
 
     @property
     def description(self) -> str:
